refactor(core): consolidate connection helper boundary
Core consolidation Slice 3. Inline the two tiny connection helper modules,
core/connection_state.py (3 state-bookkeeping helpers) and core/disconnect.py
(close_client_connection, disconnect_locked), into core/connection_lifecycle.py
with byte-identical function bodies and merged imports. The three connection-helper
files become one; core/ file count drops from 23 to 21.

The sole caller, core/client_connection.py, now imports mark_connection_established,
mark_connection_failure, mark_connection_disconnected, close_client_connection,
disconnect_locked and ensure_connected_lifecycle all from connection_lifecycle.
No re-export shims were added (per CLAUDE.md). connection_lifecycle.py has no
intra-core imports, so no circular import is introduced.

Import sites updated directly:
- core/client_connection.py
- tests/test_coordinator_connection_state.py (was core.connection_state)
- tests/test_coordinator_disconnect.py (was core.disconnect)

This is a pure file move with zero runtime behavior change. Read path, write path,
scanner, and config flow are untouched. No Modbus register addresses/names, entity
IDs, unique IDs, service IDs, or translation keys changed.

Docs: docs/core_consolidation_plan.md marks Slice 3 DONE (with rollback plan and a
correction to the readiness note, which had wrongly stated no tests imported the
modules); docs/architecture/file_inventory.md drops disconnect.py; CHANGELOG.md
[Unreleased] Internal entry added.

The maintainer explicitly waived the plan's real-device-validation gate for this
zero-behavior-change move; all other Slice 3 preconditions were already met.
