+++ custom_components/thessla_green_modbus/core/connection_lifecycle.py
@@ -1,11 +1,71 @@
-"""Connection lifecycle orchestration for device-client runtime."""
+"""Connection lifecycle, state, and disconnect helpers for device-client runtime."""
 
 from __future__ import annotations
 
-from collections.abc import Callable
+import inspect
+import logging
+from collections.abc import Callable, MutableMapping
 from typing import Any
 
 from pymodbus.exceptions import ConnectionException, ModbusException
+
+
+def mark_connection_established(*, offline_state_setter: Any) -> None:
+    """Mark coordinator runtime as connected."""
+    offline_state_setter(False)
+
+
+def mark_connection_failure(
+    *, statistics: MutableMapping[str, Any], offline_state_setter: Any
+) -> None:
+    """Record connection failure counters and switch coordinator offline."""
+    statistics["connection_errors"] += 1
+    offline_state_setter(True)
+
+
+def mark_connection_disconnected(*, offline_state_setter: Any) -> None:
+    """Mark coordinator runtime as disconnected."""
+    offline_state_setter(True)
+
+
+async def close_client_connection(*, client: Any, logger: logging.Logger) -> None:
+    """Close client object safely for sync or async close implementations."""
+    try:
+        close = getattr(client, "close", None)
+        if callable(close):
+            if inspect.iscoroutinefunction(close):
+                await close()
+            else:
+                result = close()
+                if inspect.isawaitable(result):
+                    await result
+    except (ModbusException, ConnectionException):
+        logger.debug("Error disconnecting", exc_info=True)
+    except OSError:
+        logger.exception("Unexpected error disconnecting")
+
+
+async def disconnect_locked(
+    *,
+    transport: Any,
+    client: Any,
+    close_client_connection_fn: Callable[..., Any],
+    mark_connection_disconnected_fn: Callable[[], None],
+    logger: logging.Logger,
+) -> None:
+    """Disconnect from Modbus device without acquiring locks."""
+    if transport is not None:
+        try:
+            await transport.close()
+        except (ModbusException, ConnectionException):
+            logger.debug("Error disconnecting", exc_info=True)
+        except OSError:
+            logger.exception("Unexpected error disconnecting")
+    elif client is not None:
+        await close_client_connection_fn(client=client, logger=logger)
+
+    mark_connection_disconnected_fn()
+    logger.debug("Disconnected from Modbus device")
 
 
 async def ensure_connected_lifecycle(
